move compoundButton from change to click
